cheri: Use uintptr_t sk_buff
For CHERI several fields in struct sk_buff must be able to hold full
capabilities. Change those to uintptr_t and adjust users. The field
"dev_scratch" can not hold a pointer. Instead the storage of the
pointer @dev pointer is used to store non-pointer data that fits into
an unsigned long.

+++ net/openvswitch/actions.c
@@ -782,7 +782,7 @@ static void ovs_fragment(struct net *net, struct vport *vport,
 
 	if (key->eth.type == htons(ETH_P_IP)) {
 		struct rtable ovs_rt = { 0 };
-		unsigned long orig_dst;
+		uintptr_t orig_dst;
 
 		prepare_frag(vport, skb, orig_network_offset,
 			     ovs_key_mac_proto(key));
@@ -797,7 +797,7 @@ static void ovs_fragment(struct net *net, struct vport *vport,
 		ip_do_fragment(net, skb->sk, skb, ovs_vport_output);
 		refdst_drop(orig_dst);
 	} else if (key->eth.type == htons(ETH_P_IPV6)) {
-		unsigned long orig_dst;
+		uintptr_t orig_dst;
 		struct rt6_info ovs_rt;
 
 		prepare_frag(vport, skb, orig_network_offset,
